move the serializers in the vouch app
