+++ backend/vouch/views.py
@@ -4,7 +4,7 @@ API endpoint to manipulate vouchers
 from drf_spectacular.utils import extend_schema, extend_schema_view
 from rest_framework import generics
 
-from tournesol.serializers.voucher import GivenVoucherSerializer, ReceivedVoucherSerializer
+from vouch.serializers import GivenVoucherSerializer, ReceivedVoucherSerializer
 from vouch.models import Voucher
